create models module
