+++ ipsuite/analysis/model/__init__.py
@@ -1,0 +1,27 @@
+from ipsuite.analysis.model.dynamics import (
+    BoxHeatUp,
+    BoxScaleAnalysis,
+    ConnectivityCheck,
+    EnergySpikeCheck,
+    MDStabilityAnalysis,
+    RattleAnalysis,
+)
+from ipsuite.analysis.model.predict import (
+    AnalyseForceAngles,
+    AnalysePrediction,
+    InterIntraForces,
+    PredictWithModel,
+)
+
+__all__ = [
+    "PredictWithModel",
+    "AnalyseForceAngles",
+    "AnalysePrediction",
+    "InterIntraForces",
+    "RattleAnalysis",
+    "BoxHeatUp",
+    "BoxScaleAnalysis",
+    "EnergySpikeCheck",
+    "ConnectivityCheck",
+    "MDStabilityAnalysis",
+]
